fix allows additional textanswers saving bug
When changing a the type from text/heading to a option that allows additional textanswers and clicking the checkbox, it would not be saved. This is due to the fact, that this form field was disabled and therefore was not send for the save.

I now replaced this with and attribut so I can be disabled in the frontend and be send for the submit.