fix display bug
since we are dispaching the change event on load, all set values (from the db) are overridden in the frontend. This is fine for Heading and Text (since they are unchecked and disabled) but for the other types we cannot set the checkbox to checked. Same applies when its a dropout questtionaire.